Refactor models and add support for batching
