+++ python-client/docs/reference/models/integrations/huggingface.rst
@@ -1,0 +1,5 @@
+HuggingFace models
+==================
+
+.. automodule:: giskard.models.huggingface
+    :members: HuggingFaceModel
