+++ python-client/giskard/models/sklearn/__init__.py
@@ -1,23 +1,15 @@
-from typing import Callable, Iterable, Any, Optional
+from typing import Any, Callable, Iterable, Optional
 
 import mlflow
 import pandas as pd
 
-from giskard.core.core import ModelType, SupportedModelTypes
-from giskard.core.validation import configured_validate_arguments
-from giskard.models.base import MLFlowBasedModel
+from ...core.core import ModelType, SupportedModelTypes
+from ...core.validation import configured_validate_arguments
+from ..base.serialization import MLFlowSerializableModel
 
 
-class SKLearnModel(MLFlowBasedModel):
-    """
-    The SKLearnModel class is a subclass of MLFlowBasedModel that wraps a scikit-learn model.
-    This class provides a way to standardize the API for scikit-learn models to make them compatible with
-    the other model types in the giskard package.
-
-    Attributes:
-        _feature_names_attr (str):
-            A private attribute that holds the name of the scikit-learn model's attribute that stores the feature names.
-    """
+class SKLearnModel(MLFlowSerializableModel):
+    """Automatically wraps sklearn models for use with Giskard."""
 
     _feature_names_attr = "feature_names_in_"
 
@@ -32,6 +24,7 @@ class SKLearnModel(MLFlowBasedModel):
         feature_names: Optional[Iterable] = None,
         classification_threshold: Optional[float] = 0.5,
         classification_labels: Optional[Iterable] = None,
+        batch_size: Optional[int] = None,
         **kwargs,
     ) -> None:
         """
@@ -52,6 +45,7 @@ class SKLearnModel(MLFlowBasedModel):
                 A float classification threshold value, if applicable to the model being used.
             classification_labels (Iterable, optional):
                 An iterable of classification label names, if applicable to the model being used.
+            batch_size (int, optional): The batch size to use for inference. Disabled by default.
         """
         if model_type == SupportedModelTypes.CLASSIFICATION:
             if classification_labels is None and hasattr(model, "classes_"):
