+++ python-client/pyproject.toml
@@ -18,7 +18,7 @@ test.env = { GSK_DISABLE_ANALYTICS = "True"}
 test-fast.cmd = "pytest -n auto -m 'not slow' -c pyproject.toml tests --cov=giskard --cov-report=xml --disable-warnings"
 test-fast.env = { GSK_DISABLE_ANALYTICS = "True"}
 lint = "ruff giskard tests"
-doc = "python -m sphinx_autobuild docs docs/_build/html"
+doc = "python -m sphinx_autobuild --watch giskard docs docs/_build/html"
 
 [tool.pdm.dev-dependencies]
 dev = [
@@ -252,7 +252,7 @@ warn_unused_ignores = false
 # https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
 # Directories that are not visited by pytest collector:
 norecursedirs = ["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
-doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]
+doctest_optionflags = ["ELLIPSIS", "NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]
 markers = [
     "slow: marks tests as slow (deselect with '-m \"not slow\"')",
 ]
